% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmtf.R
\name{about.cnmtf}
\alias{about.cnmtf}
\title{Read me first!!}
\usage{
about.cnmtf()
}
\description{
cNMTF: Prioritisation of single nucleotide variants using Corrected non-negative matrix tri-factorisation

A data fusion framework for prioritising reliable associations between single nucleotide variants (SNVs) and traits.
This algorithm allows for studying the effect of SNVs on categorical traits, thanks to its main features : 1) It captures the interrelatedness between variants data, the SNVs deleteriousness effect and the protein-protein interactions (PPIs) that might be disrupted.  2) It simultaneously accounts for the patient's outcome and ancestry by means of kernels functions, minimizing the confounding for population structures.

The \strong{cnmtf} package provides four categories of functions for
preprocessing data, clustering, scoring SNVs and comparing results.
}
\section{Preprocessing functions}{

These functions will help you to create the inputs for the algorithm.
}

\section{Factorisation functions}{

Main functions to integrate the input data, generate the low-dimmensional matrices and find consensus clusters.
}

\section{Scoring functions}{

A set of functions to score SNVs and prioritise significant SNV-trait associations from the low-dimmensional matrices.
}

\section{Comparing functions}{

Auxiliary functions to compare your results across different settings of the algorithm.
}

\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
