% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_main.R
\name{analyze.cnmtf}
\alias{analyze.cnmtf}
\title{Analyze cNMTF results}
\usage{
analyze.cnmtf(trait.project = NULL, name.exp = NULL, work.dat = NULL,
  alpha.cnmtf = NULL, d.conf = NULL, snps.known = NULL,
  snps.known2 = NULL, tmap = NULL)
}
\arguments{
\item{trait.project}{Trait}

\item{name.exp}{Name of experiment to analyse}

\item{work.dat}{Working directory}

\item{alpha.cnmtf}{Significance level for the delta SNV score}
}
\value{
Plots of delta scores and workspaces with results printed to files
}
\description{
Main function to perform analysis of cNMTF results
}
\seealso{
Other Scoring functions: \code{\link{custom.venn}},
  \code{\link{delta.score}}, \code{\link{manhattan.plot}},
  \code{\link{manhattan.table}},
  \code{\link{plot.clusters}}, \code{\link{rec.list}},
  \code{\link{sd.plot}}, \code{\link{sd.score}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
