% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_res.R
\name{annotate.results}
\alias{annotate.results}
\title{Annotate prioritised SNVs}
\usage{
annotate.results(name.exp = NULL, snps.known = NULL, snps.known2 = NULL,
  add.david.annotations = TRUE, email.david = NULL,
  add.ensemble.conseq = FALSE, work.dat = NULL, tmap = NULL,
  file.LD = NULL, ld.tao = 0.8)
}
\arguments{
\item{snps.known2}{List of known associations}

\item{add.david.annotations}{Logical. Use DAVID web service or export/import manually}

\item{email.david}{Email account registered in DAVID.}

\item{lconsortiums}{List of data consortiums to merge}

\item{ltrait.project}{List of traits to merge}

\item{explore.gwas}{Logical. Explore mutations from GWAS cataloge}
}
\value{
\itemize{
\item \code{cdav.all} Chart of annotations from DAVID. Printed to file "summary_david_chart.txt"
\item \code{tsa.all} Table of annotations from DAVID. Printed to file "summary_david_table.txt"
\item \code{tmut.all} Table merging annotations for a list of consortiums and traits. Printed to file "summary_david_table_genes.txt"
\item \code{tres} Final table with metadata for all SNVs. Printed to file "explore_results.csv"
}
}
\description{
Function to map SNVs to genes and retrieve functional annotations from DAVID
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
