% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_cons.R
\name{clus.membership}
\alias{clus.membership}
\title{Calculate connectivity matrix}
\usage{
clus.membership(H)
}
\arguments{
\item{H}{Factorization matrix}

\item{indCluster}{Clustering assignemnt vector}
}
\value{
\code{C} Binary connectivity matrix
}
\description{
Function to extract clustering assignments from a factorization matrix
}
\examples{
clus.membership(V)

}
\seealso{
Other Factorisation functions: \code{\link{cnmtf}},
  \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{regression.snps}}, \code{\link{score.cnmtf}},
  \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
