% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_clust.R
\name{cnmtf}
\alias{cnmtf}
\title{cNMTF algorithm}
\usage{
cnmtf(R, HLH, Vo, Wu, k, iters = 1, calcObj, calcObj2, displ = TRUE,
  tof = 1e-05, lparameters, init = 1, V.init = NULL, U.init = NULL,
  init.S = FALSE, DWD = NULL)
}
\arguments{
\item{R}{Data matrix, \emph{n x m}}

\item{HLH}{Term in the correction for population structures. Dot product: H . L . H where: H is the Centering matrix \emph{(m x m)} and L is Ancestry matrix (i.e. Kernel on the confounder random variable) \emph{(m x m)}}

\item{Vo}{Phenotype/Outcome label matrix, \emph{m x k2}}

\item{Wu}{Adjacency matrix of the SNV-SNV network. \emph{n x n}}

\item{k}{Vector of rank parameters, \emph{k1 x k2}}

\item{iters}{Default number of itersatiors}

\item{calcObj}{Check convergency each X number of itersations}

\item{displ}{Logical. Print number of iterations}

\item{tof}{RelativeError}

\item{lparameters}{Vector of regularization parameters:  \eqn{ \gamma_{1}, \gamma_{2}, \gamma_{3} } for the SNV-SNV network, the Phenotype matrix and the Ancestry Kernel respectively.}

\item{init}{Initialize the matrices randomly (0) or by using PSVD (1)}

\item{DWD}{Term in the normalised graph laplacian. Dot product:  1/sqrt(Du) . Wu . 1/sqrt(Du), where: Du is the Degree matrix and Wu is the adjacency matrix of the SNV-SNV network.}
}
\value{
\itemize{
\item \code{U} low-dimensional cluster indicator matrix for features observed data, \emph{n x k1}
\item \code{V} low-dimensional cluster indicator matrix for observed data, \emph{m x k2}
\item \code{S} cluster mapping matrix, \emph{k1 x k2}
\item \code{num.iters} number of itersation till convergence
\item \code{final.J} Final objective function value
\item \code{objF.vector} Objective function values across iterations
}
}
\description{
Function to perform one repetition of cNMTF
}
\references{
\itemize{
\item Shang, Fanha, et al. "Graph dual regularization non-negative matrix factorization for co-clustering" Pattern Recognition 45 (2012)
\item Li, Rakitisch, et al. "ccSVM: correcting SVMs for confounding factors in biological data classification" ISMB 2011
}
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{regression.snps}}, \code{\link{score.cnmtf}},
  \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
