% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_snv.R
\name{coding.scheme}
\alias{coding.scheme}
\title{Select genetic model for the genotyping data}
\usage{
coding.scheme(R, coding = c("recessive", "dominant"))
}
\arguments{
\item{R}{Relationship matrix}

\item{coding}{Desired coding scheme (genetic model) for the entries of \code{R}}
}
\value{
Relationship matrix \code{R} with a given coding scheme
}
\description{
Function to change coding scheme in a relationship matrix
}
\seealso{
Other Preprocessing functions: \code{\link{construct.Vo}},
  \code{\link{construct.Wu}}, \code{\link{create.network}},
  \code{\link{degree.distribution}},
  \code{\link{kernels.cnmtf}},
  \code{\link{mapply.snpt.synthetic.gwas}},
  \code{\link{rownet.hub}},
  \code{\link{snpt.synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
