% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_out.R
\name{construct.Vo}
\alias{construct.Vo}
\title{Constructs the outcome matrix}
\usage{
construct.Vo(out, ki)
}
\arguments{
\item{out}{Outcome vector for the patients}

\item{ki}{Number of clusters/columns for the outcome matrix}
}
\value{
\code{Vo} outcome matrix of size \link{m x ki}
}
\description{
Function to construct \code{Vo} (Low-rank matrix of outcome labels)
}
\seealso{
Other Preprocessing functions: \code{\link{coding.scheme}},
  \code{\link{construct.Wu}}, \code{\link{create.network}},
  \code{\link{degree.distribution}},
  \code{\link{kernels.cnmtf}},
  \code{\link{mapply.snpt.synthetic.gwas}},
  \code{\link{rownet}}, \code{\link{seed.snps}},
  \code{\link{snpt.synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
