% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_net.R
\name{construct.Wu}
\alias{construct.Wu}
\title{Constructs the SNV-SNV adjacency matrix}
\usage{
construct.Wu(R.snps = NULL, work.dat = NULL, trait.project = NULL,
  dedges = NULL, n.cores = NULL, net.type, tmap = NULL, lhubs = NULL,
  alledges.snp = FALSE)
}
\arguments{
\item{R.snps}{list of SNPs}

\item{trait.project}{project's name to be included in the file name}

\item{tmap}{mapping of SNPs to gene entrezids}

\item{lhubs}{list of damaging variants}

\item{ncores}{number of cores for parallel computing}
}
\value{
\code{Wu}, \code{G} adjacency matrix and graph object of the network
}
\description{
Function to construct the adjacency matrix Wu for the SNV-SNV network
}
\seealso{
Other Preprocessing functions: \code{\link{coding.scheme}},
  \code{\link{construct.Vo}}, \code{\link{create.network}},
  \code{\link{degree.distribution}},
  \code{\link{kernels.cnmtf}},
  \code{\link{mapply.snpt.synthetic.gwas}},
  \code{\link{rownet.hub}},
  \code{\link{snpt.synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
