% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_net.R
\name{create.network}
\alias{create.network}
\title{Constructs the SNV-SNV network}
\usage{
create.network(net.type = "ppi", file.dedges = NULL, remove.highLD = TRUE,
  ld.tao = 0.8, res.ld = NULL, keep.with.LD = NULL, R.snps = NULL,
  work.dat = NULL, trait.project = NULL, n.cores = 4, tmap = NULL,
  alledges.snp = FALSE, plot.file = NULL, venn.diag = FALSE)
}
\arguments{
\item{net.type}{Type of reference network. Default "ppi"}

\item{file.dedges}{Filename with edges from reference network}

\item{ld.tao}{Treshold of LD. Default 0.8}

\item{res.ld}{Table of LD}

\item{keep.with.LD}{List of SNPs to include even if they are in LD}

\item{R.snps}{List of SNPs in R}

\item{work.dat}{Working directory}

\item{trait.project}{Trait}

\item{n.cores}{Number of cores for parallel computing}

\item{tmap}{Mapping of SNPs to genes}

\item{alledges.snp}{Include edges between proteins for all SNPs or only for hubs}

\item{plot.file}{File to print Venn diagrams and node degree distribution}

\item{venn.diag}{Logical. Print Venn diagrams. Default = FALSE.}

\item{expand.highLD}{Expand SNV consequences to SNVs in high LD}
}
\value{
\itemize{
\item \code{Wu}, \code{G}: adjacency matrix and graph object of the network
\item Table with node properties to be used in Cytoscape
\item Venn diagrams of damaging variants and node degree distribution
}
}
\description{
Function to create a SNV-SNV network using a PPI reference network
}
\seealso{
Other Preprocessing functions: \code{\link{coding.scheme}},
  \code{\link{construct.Vo}}, \code{\link{construct.Wu}},
  \code{\link{degree.distribution}},
  \code{\link{kernels.cnmtf}},
  \code{\link{mapply.snpt.synthetic.gwas}},
  \code{\link{rownet}}, \code{\link{seed.snps}},
  \code{\link{snpt.synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
