% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_plot.R
\name{custom.venn}
\alias{custom.venn}
\title{Plot Venn Diagrams}
\usage{
custom.venn(venn.list, main = NULL, filename = NULL)
}
\arguments{
\item{venn.list}{list of elements in each set}

\item{main}{title for the diagram}

\item{fill.venn}{colour for set}
}
\value{
Printed venn diagram to console
}
\description{
Function to plot customised the venn diagrams
}
\seealso{
Other Scoring functions: \code{\link{analyze.cnmtf}},
  \code{\link{delta.score}}, \code{\link{manhattan.plot}},
  \code{\link{manhattan.table}},
  \code{\link{plot.clusters}}, \code{\link{rec.list}},
  \code{\link{sd.plot}}, \code{\link{sd.score}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
