% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_net.R
\name{degree.distribution}
\alias{degree.distribution}
\title{Plot node degree of a network}
\usage{
degree.distribution(Gu, gamma.kd, weighted = FALSE)
}
\arguments{
\item{Gu}{graph object}

\item{gamma.kd}{gamma parameter for a law power distribution of node degree}

\item{weighted}{are the edges weighted?}
}
\value{
Print graph variables and plot node degree distribution
}
\description{
Function to plot node degree distribution and calculate some graph variables
}
\seealso{
Other Preprocessing functions: \code{\link{coding.scheme}},
  \code{\link{construct.Vo}}, \code{\link{construct.Wu}},
  \code{\link{create.network}},
  \code{\link{kernels.cnmtf}},
  \code{\link{mapply.snpt.synthetic.gwas}},
  \code{\link{rownet.hub}},
  \code{\link{snpt.synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
