% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_aux.R
\name{delta.score}
\alias{delta.score}
\title{Delta score between clusters}
\usage{
delta.score(file.exp1, file.exp2 = NULL, snps.known1, snps.known2,
  print.file = NULL, sig.snp.nmtf = NULL, clus.a = list(c(1)),
  clus.b = list(c(2)))
}
\arguments{
\item{file.exp1}{workspace of the experiment (created with function score.cnmtf)}

\item{file.exp2}{Optional. Workspace of a second experiment.}

\item{snps.known1}{List of known associations}

\item{snps.known2}{Second list of known associations}

\item{print.file}{File to print plots}

\item{sig.snp.nmtf}{List of significant SNVs in this experiment}

\item{clus.a, clus.b}{List of clusters of patients to compute delta scores. Clusters in \code{clus.a} are compared with clusters in \code{clus.b}.}

\item{snps.known}{List of known associations}
}
\value{
\code{dOn} Delta Standardised Omega matrix with the delta scores for each pair of clusters.
}
\description{
Auxiliar function to calculate delta score between clusters of patients.
}
\seealso{
Other Scoring functions: \code{\link{analyze.cnmtf}},
  \code{\link{custom.venn}}, \code{\link{manhattan.plot}},
  \code{\link{manhattan.table}},
  \code{\link{plot.clusters}}, \code{\link{rec.list}},
  \code{\link{sd.plot}}, \code{\link{sd.score}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
