% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_cons.R
\name{hierarchical.clust}
\alias{hierarchical.clust}
\title{Hierarchical clustering from consensus connectivity matrices}
\usage{
hierarchical.clust(aC, ki, labels, linkF = c("average"))
}
\arguments{
\item{aC}{consensus matrix}

\item{labels}{sample labels}

\item{linkF}{linkage function ('average','centroid','complete', etc.)}

\item{k}{rank parameter (number of clusters)}

\item{out_filename}{exporting label assignment to clusters}
}
\value{
\code{res} cluster assignment
}
\description{
Function for extracting cluster labels based on hierarchical clustering
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{cnmtf}}, \code{\link{consensus.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{regression.snps}}, \code{\link{score.cnmtf}},
  \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
