% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_init.R
\name{initialise.UV}
\alias{initialise.UV}
\title{Initialising the low-rank matrices U and V}
\usage{
initialise.UV(R, name.init = NULL, work.dat = NULL, k2, k1,
  name.exp = NULL, log.file = NULL)
}
\arguments{
\item{R}{Relationship matrix, \emph{n x m}}

\item{name.init}{Name of the workspace with initialisations of U and V}

\item{work.dat}{Folder to read or write a workspace with initialisations}

\item{k1, k2}{Rank parameters (number of clusters for U and V, respectively)}

\item{logfile}{Name of a file to write progess/status of this function}
}
\value{
\code{U.init}, \code{V.init} intialisations of U and V.
}
\description{
This function creates initialisations of U and V. It can also read stored initialisations to speed up the cNMTF algorithm.
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{cnmtf}}, \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{regression.snps}}, \code{\link{score.cnmtf}},
  \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
