% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_plot.R
\name{manhattan.plot}
\alias{manhattan.plot}
\title{Manhattan plot}
\usage{
manhattan.plot(pvalues, lchr = 1:22, color, pos.snps.known1 = NULL,
  pos.snps.known2 = NULL, print.file = NULL, alpha = 0.05, med.point)
}
\arguments{
\item{pvalues}{pvalues of the SNPs in tmanhattan}

\item{lchr}{List of chromosomes}

\item{color}{colour for each SNP}

\item{pos.snps.known1, pos.snps.known2}{Positions of known associations}

\item{print.file}{PDF file to print plots}

\item{alpha}{Significance level}

\item{med.point}{parameter to place the axes labels in a manhattan plot}
}
\value{
Print manhattan plot either in console or in print.file
}
\description{
Function to plot a Manhattan plot
}
\seealso{
Other Scoring functions: \code{\link{analyze.cnmtf}},
  \code{\link{custom.venn}}, \code{\link{delta.score}},
  \code{\link{manhattan.table}},
  \code{\link{plot.clusters}}, \code{\link{rec.list}},
  \code{\link{sd.plot}}, \code{\link{sd.score}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
