% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_plot.R
\name{manhattan.table}
\alias{manhattan.table}
\title{Manhattan table}
\usage{
manhattan.table(l.snps, tmap)
}
\arguments{
\item{l.snps}{List of SNPs to construct the table}

\item{tmap}{Mapping of SNPs to genes, chr and genomic position}
}
\value{
\itemize{
\item tmanhattan : table of chromosome position of SNPs.
\item med.point : parameter to place the axes labels in a manhattan plot
}
}
\description{
Function to create a table of chromosomal position used in manhatan plots
}
\seealso{
Other Scoring functions: \code{\link{analyze.cnmtf}},
  \code{\link{custom.venn}}, \code{\link{delta.score}},
  \code{\link{manhattan.plot}},
  \code{\link{plot.clusters}}, \code{\link{rec.list}},
  \code{\link{sd.plot}}, \code{\link{sd.score}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
