% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_syn.R
\name{mapply.snpt.synthetic.gwas}
\alias{mapply.snpt.synthetic.gwas}
\title{Simulate synthetic genotypes}
\usage{
mapply.snpt.synthetic.gwas(comb.snpt, comb.snpn)
}
\arguments{
\item{comb.snpt}{Combination of SNP types in the synthetic dataset}

\item{comb.snpn}{Number of SNP types}
}
\value{
\itemize{
\item \code{R} synthetic GWAS data set
\item \code{out} disease status of the patient
\item \code{pop} population ancestry of the patient
\item \code{srisk} genotype relative risk for each SNP
\item \code{snp.type} SNP type
\item \code{snp.cat} SNP category (Association with the disease)
}
}
\description{
Function to generate synthetic data for a combination of SNP types and number of SNPs
}
\seealso{
Other Preprocessing functions: \code{\link{coding.scheme}},
  \code{\link{construct.Vo}}, \code{\link{construct.Wu}},
  \code{\link{create.network}},
  \code{\link{degree.distribution}},
  \code{\link{kernels.cnmtf}}, \code{\link{rownet.hub}},
  \code{\link{snpt.synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
