% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_param.R
\name{parameters.cnmtf}
\alias{parameters.cnmtf}
\title{Optimal penalization paramaters}
\usage{
parameters.cnmtf(R, out.cat = NULL, pop = NULL, k, log.file = NULL,
  Wv = NULL, Wu = NULL, HLH = NULL, Vo = NULL, Gu = NULL, Gv = NULL,
  V.init = NULL, U.init = NULL, name.exp = NULL, name.init = NULL,
  work.dat = NULL, sequential.estimation = FALSE, wparameters = NULL,
  run.t.par = NULL, range.parameters = NULL, max.try0 = NULL,
  calcObj = NULL, calcObj2 = NULL, init = NULL, parallel.opt = FALSE,
  n.cores = 2, iters = NULL, display.iters = TRUE, tao.sc = 4,
  snps.known = NULL)
}
\arguments{
\item{R}{Relationship matrix}

\item{out.cat}{Categorical outcome variable}

\item{pop}{Population variable}

\item{k}{Number of clusters}

\item{log.file}{Log file to track status of the function}

\item{Wv, Wu}{Adjacency matrices to guide clustering of U and V}

\item{HLH}{Term to correct for population stratification}

\item{Gu}{iGraph for Wu (SNP-SNP network)}

\item{Gv}{iGraph for Wv (Patient-Patient network)}

\item{U.init, V.init}{Initialisations for matrices U and V}

\item{name.exp}{Name of experiment to save files}

\item{name.init}{Name of workspace with initialisations of U and V}

\item{work.dat}{Folder to save and load workspaces}

\item{sequential.estimation}{Set the parameters in a specific order and carry the optimals}

\item{wparameters}{Weights of penalisation terms to be computed}

\item{run.t.par}{Number of repetitions for parameters fitting}

\item{range.parameters}{Range of parameters to be evaluated (if estimate.par == FALSE)}

\item{max.try0}{Maximum number of tries to fit the parameter}

\item{calcObj, calcObj2}{Check convergency each X number of iterations after first Y iterations}

\item{init}{Type of seeding/initialisation of matrices in the algorithm}

\item{parallel.opt}{Run some instances of the algorithm in parallel}

\item{n.cores}{Number of cores to use in the parallel processing}

\item{iters}{Number of iterations}

\item{display.iters}{Display the iterations of function cnmtf}

\item{tao.sc}{Trehsold of standard deviations for the SNV score}

\item{snps.known}{List of known associations}
}
\value{
\code{opt.parameters} list of optimal parameters fitted: \code{gamma1, gamma2, gamma3}
}
\description{
Function to estimate the optimal parameters for the cNMTF algorithm.
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{cnmtf}}, \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{regression.snps}}, \code{\link{score.cnmtf}},
  \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
