% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_plot.R
\name{plot.clusters}
\alias{plot.clusters}
\title{Plot cNMTF clusters}
\usage{
\method{plot}{clusters}(file.exp, snps.known, print.file = NULL, maf.snps.cat,
  R.snps)
}
\arguments{
\item{print.file}{File to print plots}

\item{maf.snps.cat}{Minor Allele Frequency}

\item{R.snps}{List of SNVs}

\item{file.exp:}{Workspace of the experiment (created with function score.cnmtf)}

\item{snps.known:}{List of known associations}
}
\value{
Plots of clusters printed to \code{print.file}
}
\description{
Function to plot heatmaps and clusters of SNPs and patients
}
\seealso{
Other Scoring functions: \code{\link{analyze.cnmtf}},
  \code{\link{custom.venn}}, \code{\link{delta.score}},
  \code{\link{manhattan.plot}},
  \code{\link{manhattan.table}}, \code{\link{rec.list}},
  \code{\link{sd.plot}}, \code{\link{sd.score}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
