% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_param.R
\name{plot.parameter}
\alias{plot.parameter}
\title{Plot penalization parameters}
\usage{
\method{plot}{parameter}(x, y, y.max, y.name, x.name, fitmodel = FALSE, ww,
  print.file = NULL, range.fit = NULL, log.file = NULL,
  logarithm.x = FALSE)
}
\arguments{
\item{x}{Vector of parameter values}

\item{y}{Tracking variable}

\item{y.max}{Maximum value of the tracking variable}

\item{y.name}{Name of tracking variable}

\item{x.name}{Name of the parameter}

\item{fitmodel}{Fit a linear model or a cubic spline on the data}

\item{ww}{Information weigth to find optimal point between y.min and y.max}

\item{print.file}{Print plots to file}

\item{range.fit}{Range of values of x to fit the model}

\item{log.file}{Log.file to print results}

\item{logarithm.x}{Plot x axis in logarithmic scale}
}
\value{
\code{opt.param.i} Vector of optimal parameters if \code{fitmodel == TRUE}, otherwise the function returns only a plot of both variables
}
\description{
Function to plot changes in parameter versus a tracking variable
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{cnmtf}}, \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{regression.snps}}, \code{\link{score.cnmtf}},
  \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
