% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_init.R
\name{psvd.init}
\alias{psvd.init}
\title{SVD Initialisation for NMF}
\usage{
psvd.init(A, k)
}
\arguments{
\item{A}{Nonnegative matrix A, \emph{n x n}}

\item{k}{Rank of the computed factors (number of clusters)}
}
\value{
\code{W} nonnegative matrix, \emph{n x k}
}
\description{
Function to initialize Non-negative Matrix Factorization algorithms
}
\references{
C. Boutsidis and E. Gallopoulos, SVD-based initialization: A head start for nonnegative matrix factorization, Pattern Recognition, Elsevier
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{cnmtf}}, \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}},
  \code{\link{regression.snps}}, \code{\link{score.cnmtf}},
  \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
