% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R}
\alias{R}
\title{Relationship matrix of SNVs and patients}
\format{The table consists of 100 rows (SNVs) and 50 columns (patients)
\describe{
  \item{Entries}{Number of copies of the recessive allele (0,1,2) at a given SNV and patient}
}}
\usage{
R
}
\description{
A data table containing the genotypes of patients (0,1,2) at each SNV.
}
\keyword{datasets}
