% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_plot.R
\name{rec.list}
\alias{rec.list}
\title{Recursive lists}
\usage{
rec.list(len)
}
\arguments{
\item{len}{vector with the length of list levels.}
}
\value{
Return a nested list with different lengths at each level
}
\description{
Function to create recursive lists
}
\seealso{
Other Scoring functions: \code{\link{analyze.cnmtf}},
  \code{\link{custom.venn}}, \code{\link{delta.score}},
  \code{\link{manhattan.plot}},
  \code{\link{manhattan.table}},
  \code{\link{plot.clusters}}, \code{\link{sd.plot}},
  \code{\link{sd.score}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
