% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_lrm.R
\name{regression.snps}
\alias{regression.snps}
\title{Prioritise variants using regression models}
\usage{
regression.snps(out, R, logistic.model = TRUE, coding = "additive",
  d.conf = NULL)
}
\arguments{
\item{out}{clinical output}

\item{R}{relationship matrix}

\item{logistic.model}{Use logistic regression or linear regression models}

\item{coding}{Codding scheme for the R matrix}

\item{d.conf}{dataframe with confounder variables to correct for}
}
\value{
\code{model.pvals} p-values for the association SNV-trait
}
\description{
Function to fit logistic or linear regression models
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{cnmtf}}, \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{score.cnmtf}}, \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
