% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_net.R
\name{rownet}
\alias{rownet}
\title{Creates edge list for the SNV-SNV network}
\usage{
rownet(i = NULL, dedges = NULL, lgene = NULL, R.snps = NULL,
  tmap = NULL)
}
\arguments{
\item{i}{counter variable over the list of genes}

\item{dedges}{network as a list of edges between gene entrezids}

\item{lgene}{list of genes mapping the SNPs in R}

\item{R.snps}{list of SNPs in the relationship matrix}

\item{tmap}{table for mapping SNPs to genes}
}
\value{
Nested list of SNP positions to be connected by edges
}
\description{
Function to find pairs of SNVs linked in the network
}
\seealso{
Other Preprocessing functions: \code{\link{coding.scheme}},
  \code{\link{construct.Vo}}, \code{\link{construct.Wu}},
  \code{\link{create.network}},
  \code{\link{degree.distribution}},
  \code{\link{kernels.cnmtf}},
  \code{\link{mapply.snpt.synthetic.gwas}},
  \code{\link{seed.snps}},
  \code{\link{snpt.synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
