% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_main.R
\name{score.cnmtf}
\alias{score.cnmtf}
\title{Calculate Omega matrix}
\usage{
score.cnmtf(R, out = NULL, pop = NULL, log.file = NULL, name.exp = NULL,
  name.init = NULL, work.dat = NULL, define.k = c("user", "NMTF", "PCA"),
  k1 = NULL, k2 = NULL, estimate.par = TRUE, wparameters = NULL,
  save.parameters = TRUE, run.t.exp = NULL, run.t.par = 4,
  range.parameters = NULL, sequential.estimation = FALSE, max.try0 = 4,
  tao.sc = 4, snps.known = NULL, calcObj = NULL, calcObj2 = NULL,
  init = NULL, parallel.opt = FALSE, n.cores = 2, iters = NULL,
  do.U = FALSE, display.iters = TRUE, score.pvalues = c(TRUE, FALSE,
  "only"), randomisations = 50, random.parallel = FALSE,
  file.Gu = file.Gu, norm.Lu = FALSE)
}
\arguments{
\item{R}{Relationship matrix}

\item{out}{Categorical outcome variable}

\item{pop}{Population variable}

\item{log.file}{Log file to track progress of the function}

\item{name.exp}{Name of experiment to save files}

\item{name.init}{Name of workspace with initialisations of U and V}

\item{work.dat}{Folder to save and load workspaces}

\item{define.k}{Method to define \code{k1}. It can be: \code{c("user","method","PCA")}}

\item{k1}{Number of SNV clusters}

\item{k2}{Number of patient clusters equals number of levels in the outcome}

\item{estimate.par}{Logical. Estimate penalisation parameters. If FALSE then provide the parameters in the argument "wparameters". If TRUE then provide preliminar weights to the parameters in the argument "wparameters".}

\item{wparameters}{Either the Penalization parameters (if \code{estimate.par == FALSE}) or Weights of penalisation terms to be computed (if estimate.par == TRUE)}

\item{save.parameters}{Logical. Save parameters to file}

\item{run.t.exp}{Number of repetitions for the experiment}

\item{run.t.par}{Number of repetitions for parameters fitting}

\item{range.parameters}{Range of parameters to be evaluated (if \code{estimate.par == FALSE})}

\item{sequential.estimation}{Set the parameters in a specific order and carry the optimals}

\item{max.try0}{Maximum number of tries to fit the parameter}

\item{tao.sc}{Trehsold of standard deviations for the SNV score}

\item{snps.known}{List of known associations}

\item{calcObj, calcObj2}{Number of iterations to check convergency. Check convergency each \code{calcObj} number of iterations after first \code{calcObj2} iterations}

\item{init}{Type of seeding/initialisation of matrices in the algorithm}

\item{parallel.opt}{Run some instances of the algorithm in parallel}

\item{n.cores}{Number of cores to use in the parallel processing}

\item{iters}{Number of iterations}

\item{do.U}{Perform clustering of SNPs}

\item{display.iters}{Display the iterations of function cnmtf}

\item{score.pvalues}{Estimate p-values for the scores \code{c(TRUE, FALSE, "only")}.}

\item{randomisations}{Number of randomisations}

\item{random.parallel}{Logical. Run the radomisations in parallel}

\item{file.Gu}{Workspace with adjancency matrix for the SNV-SNV network.}
}
\value{
The function internally generates:
\itemize{
\item Estimation of number of SNP clusters (k1)
\item Estimation of penalisation parameters (gamma1, gamma3, gamma2)
}

The function prints the following objects in workspaces:
\itemize{
\item Initialisation of matrices U.init and V.init
\item \code{res.cnmtf}: results of factorisations
\item \code{lcnmtf.ran}: results of algorithm applied on randomisations of R
}
}
\description{
Function to calculate the Omega matrix (SNV scores per cluster of patients)
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{cnmtf}}, \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{regression.snps}},
  \code{\link{synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
