% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_plot.R
\name{sd.plot}
\alias{sd.plot}
\title{Plotting delta scores}
\usage{
sd.plot(R.snps = rownames(R), object.sd = NULL, work.dat = NULL,
  pvals.lrm = NULL, alpha.lrm = NULL, snps.known1 = NULL,
  snps.known2 = NULL, maf.snps = NULL, kd.snps = NULL,
  snps.fp.lrm = NULL, print.file = NULL, tao.sd = NULL,
  clus.a = list(c(1)), clus.b = list(c(2)), trait.project = NULL,
  tmap = NULL, ylim.sd = c(-10, 10))
}
\arguments{
\item{R.snps}{List of SNPs in the original seed files}

\item{object.sd}{Results from function sd.score}

\item{work.dat}{Working directory}

\item{pvals.lrm}{P-values obtained in a LRM}

\item{alpha.lrm}{Significance level}

\item{snps.known1}{List of known associations}

\item{snps.known2}{List of known associations 2}

\item{maf.snps}{Minor allele frequencies of SNPs in the original R matrix without filtering}

\item{kd.snps}{Node degress of SNPs in the original Wu matrix without filtering}

\item{snps.fp.lrm}{Potential False Positive association from LRM}

\item{print.file}{File to print plots}

\item{tao.sd}{Treshold of SD}

\item{clus.a}{List of patient clusters to use in the delta scores}

\item{trait.project}{Trait/outcome}

\item{tmap}{Mapping of SNPs to genes, chr and genomic position}

\item{ylim.sd}{Limits for \code{y} axis}
}
\value{
\itemize{
\item Plots printed to \code{print.file}
\item \code{sig.snp.nmtf, sig.snp.lrm} : Significant SNV ids from cNMTF and LRM
}
}
\description{
Function to plot dispersion plots, manhattan plots and venn diagrams of SNP sds
}
\seealso{
Other Scoring functions: \code{\link{analyze.cnmtf}},
  \code{\link{custom.venn}}, \code{\link{delta.score}},
  \code{\link{manhattan.plot}},
  \code{\link{manhattan.table}},
  \code{\link{plot.clusters}}, \code{\link{rec.list}},
  \code{\link{sd.score}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
