% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore_aux.R
\name{sd.score}
\alias{sd.score}
\title{Calculate standardised delta score}
\usage{
sd.score(file.exp, path.exp, name.exp, snps.known1, snps.known2,
  out.cat = NULL, pop = NULL, clus.a = list(c(1)), clus.b = list(c(2)),
  use.randomisations = TRUE, file.ran = lfile.ran[[i]],
  alpha.cnmtf = 0.005)
}
\arguments{
\item{file.exp}{Workspace of the experiment (created with function score.cnmtf)}

\item{path.exp}{Path to files of experiment}

\item{name.exp}{Name of experiment (trait and id.exp)}

\item{snps.known1}{List of known associations}

\item{snps.known2}{Second List of known associations}

\item{out.cat}{Outcome vector}

\item{pop}{Population vector}

\item{clus.a}{List of patient clusters to use in the delta scores}

\item{use.randomisations}{Logical. Use randomisations to find the optimal cutoff points for the delta scores}

\item{file.ran}{Path to file of randomisations}

\item{alpha.cnmtf}{Level of significance}
}
\value{
\itemize{
\item \code{ldOn} : list of delta scores per combination of clusters
\item \code{lsd.sc} : list of SD per combination of clusters
\item \code{tplot.sc} : table with frequencies of predicted associations
}
}
\description{
Function to extract results from cNMTF and calculate delta scores
}
\seealso{
Other Scoring functions: \code{\link{analyze.cnmtf}},
  \code{\link{custom.venn}}, \code{\link{delta.score}},
  \code{\link{manhattan.plot}},
  \code{\link{manhattan.table}},
  \code{\link{plot.clusters}}, \code{\link{rec.list}},
  \code{\link{sd.plot}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
