% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_snv.R
\name{seed.snps}
\alias{seed.snps}
\title{Select sample of SNVs}
\usage{
seed.snps(seed.file.main = NULL, seed.file.sab = NULL,
  seed.file.tal = NULL, ppi.file = NULL, lauthors = NULL,
  seed.file.results = NULL, data.consortium = NULL,
  add.random.snps = TRUE, utrait = NULL, trait = NULL)
}
\arguments{
\item{seed.file.main}{GWAS catalogue file}

\item{seed.file.sab}{Table with SNPs found by Sabatti et al}

\item{ppi.file}{PPIN from BIOGRID}

\item{lauthors}{List of authors from specific studiess}

\item{seed.file.sab}{Table with SNPs found by Talmut et al.}

\item{log.file}{File to save counts of SNPs}
}
\value{
A workspace named "\code{seed.file.results}" with the following objects:
\itemize{
\item \code{lgenes.exp.attr} Expanded list of genes from PPIN
\item \code{dseed.s} Depurated list of SNPs
\item \code{tmap.o} Table to map seed SNPs to genes
\item \code{lgenes.g.id} List of genes mapped from seed SNPs
}
}
\description{
Function to retrieve list of seed SNVs and expanded list of genes
}
\seealso{
Other Preprocessing functions: \code{\link{coding.scheme}},
  \code{\link{construct.Vo}}, \code{\link{construct.Wu}},
  \code{\link{create.network}},
  \code{\link{degree.distribution}},
  \code{\link{kernels.cnmtf}},
  \code{\link{mapply.snpt.synthetic.gwas}},
  \code{\link{rownet}}, \code{\link{snpt.synthetic.gwas}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
