% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_syn.R
\name{snpt.synthetic.gwas}
\alias{snpt.synthetic.gwas}
\title{Simulate synthetic genotypes}
\usage{
snpt.synthetic.gwas(snpt, n.snps, n.pats = 2000, n.pops = 2,
  p.pops = c(0.5, 0.5), c.pops = c(0.6, 0.4), r.pops1 = c(1, 1),
  r.pops2 = c(3, 3), pl.pop1 = c(0.1, 0.1), pl.pop2 = c(0.15, 0.45),
  fst = 0.001)
}
\arguments{
\item{snpt}{SNP type}

\item{n.snps}{number of SNPs in the synthetic dataset}

\item{n.pats}{number of patients in the synthetic dataset}

\item{n.pops}{number of patient populations in the synthetic dataset}

\item{p.pops}{proportion of patients in each population}

\item{c.pops}{proportion of cases in each patient population}

\item{fst}{Factor of differentiation between populations if \code{ allele.freq.model == "Balding-Nichols" }.}
}
\value{
\itemize{
\item \code{db.g} synthetic GWAS data set
\item \code{status} disease status of the patient
\item \code{pop} population ancestry of the patient
\item \code{snp.risk} genotype relative risk for each SNP
}
}
\description{
Function to create sythentic gwas dataset for a given snp type
}
\seealso{
Other Preprocessing functions: \code{\link{coding.scheme}},
  \code{\link{construct.Vo}}, \code{\link{construct.Wu}},
  \code{\link{create.network}},
  \code{\link{degree.distribution}},
  \code{\link{kernels.cnmtf}},
  \code{\link{mapply.snpt.synthetic.gwas}},
  \code{\link{rownet}}, \code{\link{seed.snps}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
