% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_syn.R
\name{synthetic.gwas}
\alias{synthetic.gwas}
\title{Simulate synthetic genotypes}
\usage{
synthetic.gwas(n.snps = 10, n.pats = 100, n.pops = 3, p.pops = c(0.25,
  0.25, 0.5), c.pops = c(0.2, 0.2, 0.2), r.pops = c(1.5, 1, 1.5),
  fst = 0.01, allele.freq.model = c("Balding-Nichols", "fitted"),
  pl.pop = c(0.1, 0.1, 0.9), prefix.snp = "snp")
}
\arguments{
\item{n.snps}{number of SNPs in the synthetic dataset}

\item{n.pats}{number of patients in the synthetic dataset}

\item{n.pops}{number of patient populations in the synthetic dataset}

\item{p.pops}{proportion of patients in each population}

\item{c.pops}{proportion of cases in each patient population}

\item{fst}{Factor of differentiation between populations if \code{ allele.freq.model == "Balding-Nichols" }.}

\item{allele.freq.model}{define the model to generate the allele frequency \code{ c("Balding-Nichols","fitted") }.}

\item{pl.pop}{vector of fitted allele frequencies in each population if \code{ allele.freq.model == "fitted" }.}

\item{prefix.snp}{prefix to name the SNP columns}

\item{r}{risk factor vector ( \emph{i.e.}, genotype relative risks) to generate the genotypes for cases/controls in each population}

\item{pcases}{proportion of cases}
}
\value{
\itemize{
\item \code{db.g} synthetic GWAS data set
\item \code{status} disease status of the patient
\item \code{pop} population ancestry of the patient
\item \code{snp.risk} genotype relative risk for each SNP
}
}
\description{
Function to create synthetic GWAS data sets
}
\references{
\itemize{
\item Mingyao Li, \emph{et al.} (2010) "Correcting population stratification in GWAS using a phylogenetic approach" Bioinformatics
\item Alkes Price, \emph{et al.} (2006) "Principal components analysis corrects for stratification in genome-wide association studies" Nature genetics
}
}
\seealso{
Other Factorisation functions: \code{\link{clus.membership}},
  \code{\link{cnmtf}}, \code{\link{consensus.clust}},
  \code{\link{hierarchical.clust}},
  \code{\link{initialise.UV}}, \code{\link{neg.constrain}},
  \code{\link{parameters.cnmtf}},
  \code{\link{plot.parameter}},
  \code{\link{pos.constrain}}, \code{\link{psvd.init}},
  \code{\link{regression.snps}}, \code{\link{score.cnmtf}}
}
\author{
Luis G. Leal, \email{lgl15@imperial.ac.uk}
}
