% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmap}
\alias{tmap}
\title{Mapping table for SNVs}
\format{A data frame with 50 rows and 3 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
}}
\usage{
tmap
}
\description{
A dataset containing a mapping of SVNs to genes and their chromosomal position.
}
\keyword{datasets}
